requirements.txt
Flask==3.0.0
Flask-MySQLdb==2.0.0
bcrypt==4.1.2
requests==2.31.0
python-dotenv==1.0.0
Flask-Session==0.5.0
mysqlclient==2.2.1
